+++ CHANGELOG.md
@@ -133,3 +133,4 @@
 - 2023-10-15: Metrics session, deals, qualifier, gp, +3
 - 2023-10-15: Err gp, qualifier, session, deals, +2
 - 2023-10-16: Builder qualifier, session, deals, gp, +3
+- 2023-10-17: Pool qualifier, deals, session, gp, +1